chore: change to bulk update
