+++ src/XPurchase/VirtoCommerce.XPurchase/Commands/UpdateWishlistItemsCommand.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace VirtoCommerce.XPurchase.Commands
+{
+    public class UpdateWishlistItemsCommand : WishlistCommand
+    {
+        public List<WishListItem> Items { get; set; } = new List<WishListItem>();
+    }
+}
