+++ src/XPurchase/VirtoCommerce.XPurchase/Schemas/InputUpdateWishlistLineItemType.cs
@@ -2,11 +2,10 @@ using GraphQL.Types;
 
 namespace VirtoCommerce.XPurchase.Schemas
 {
-    public class InputUpdateWishlistItemType : InputObjectGraphType
+    public class InputUpdateWishlistLineItemType : InputObjectGraphType
     {
-        public InputUpdateWishlistItemType()
+        public InputUpdateWishlistLineItemType()
         {
-            Field<NonNullGraphType<StringGraphType>>("listId", description: "Wish list id");
             Field<NonNullGraphType<StringGraphType>>("lineItemId", description: "Line Item Id to update");
             Field<NonNullGraphType<IntGraphType>>("quantity", description: "Product quantity to add");
         }
